microservice1/pyproject.toml
[tool.poetry]
name = "microservice1"
version = "0.1.0"
description = ""
authors = ["fmelihh <furkanmelihercan.98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.1"
uvicorn = "^0.30.6"
opentelemetry-distro = "^0.48b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
